Fix collab tests
Needed to do a bit of musical chairs so `title_bar` doesn't depend on `collab_ui`.

It seems that tests fail with the prevoius approach of `zed::main` directly calling `title_bar::init`.

Moving `title_bar::init` inside `collab_ui::init` seems to fix things.
