+++ crates/collab_ui/Cargo.toml
@@ -60,6 +60,7 @@ story = { workspace = true, optional = true }
 theme.workspace = true
 time_format.workspace = true
 time.workspace = true
+title_bar.workspace = true
 ui.workspace = true
 util.workspace = true
 vcs_menu.workspace = true
